# backend/docs/swagger.yaml
definitions:
  entities.Book:
    properties:
      author:
        maxLength: 255
        minLength: 2
        type: string
      cover_image_url:
        type: string
      description:
        maxLength: 1000
        type: string
      id:
        type: integer
      isbn:
        type: string
      number_of_pages:
        type: integer
      publication_date:
        type: string
      title:
        maxLength: 255
        minLength: 2
        type: string
    required:
    - author
    - isbn
    - number_of_pages
    - publication_date
    - title
    type: object
info:
  contact:
    email: bambang.handoko12@gmail.com
    name: Bambang Handoko (Ando)
  description: This is a sample API for managing books in the library.
  title: Mini Books Library API
  version: "1.0"
paths:
  /books:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all books in the library
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.Book'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all books
      tags:
      - books
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Add a new book to the library
      parameters:
      - description: Book Title
        in: formData
        name: title
        required: true
        type: string
      - description: Book Author
        in: formData
        name: author
        required: true
        type: string
      - description: Cover Image URL
        in: formData
        name: cover_image_url
        type: string
      - description: Description
        in: formData
        name: description
        type: string
      - description: Publication Date (YYYY-MM-DD)
        in: formData
        name: publication_date
        required: true
        type: string
      - description: Number of Pages
        in: formData
        name: number_of_pages
        required: true
        type: integer
      - description: ISBN (13 digits)
        in: formData
        name: isbn
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.Book'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Add a new book
      tags:
      - books
  /books/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a book by its ID
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Delete a book by ID
      tags:
      - books
    get:
      consumes:
      - application/json
      description: Get detailed information about a book by its ID
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Book'
        "404":
          description: Book not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get book by ID
      tags:
      - books
    put:
      consumes:
      - application/x-www-form-urlencoded
      description: Partially update a book's details by its ID (only provided fields
        will be updated)
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      - description: Book Title
        in: formData
        name: title
        type: string
      - description: Book Author
        in: formData
        name: author
        type: string
      - description: Cover Image URL
        in: formData
        name: cover_image_url
        type: string
      - description: Description
        in: formData
        name: description
        type: string
      - description: Publication Date (YYYY-MM-DD)
        in: formData
        name: publication_date
        type: string
      - description: Number of Pages
        in: formData
        name: number_of_pages
        type: integer
      - description: ISBN (13 digits)
        in: formData
        name: isbn
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Book'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update a book by ID
      tags:
      - books
swagger: "2.0"